+++ packages/brick_sqlite_generators/test/sqlite_schema/test_sqlite_column_type.dart
@@ -10,19 +10,16 @@ import 'package:brick_sqlite_abstract/db.dart';
 import 'package:brick_sqlite_abstract/db.dart' show Migratable;
 
 /// All intelligently-generated migrations from all `@Migratable` classes on disk
-final Set<Migration> migrations = Set.from([]);
+final Set<Migration> migrations = <Migration>{};
 
 /// A consumable database structure including the latest generated migration.
-final schema = Schema(0,
-    generatorVersion: 1,
-    tables: Set<SchemaTable>.from([
-      SchemaTable('ExplicitColumnType',
-          columns: Set.from([
-            SchemaColumn('_brick_id', int,
-                autoincrement: true, nullable: false, isPrimaryKey: true),
-            SchemaColumn('image', null, columnType: Column.blob)
-          ]))
-    ]));
+final schema = Schema(0, generatorVersion: 1, tables: <SchemaTable>{
+  SchemaTable('ExplicitColumnType', columns: <SchemaColumn>{
+    SchemaColumn('_brick_id', int,
+        autoincrement: true, nullable: false, isPrimaryKey: true),
+    SchemaColumn('image', null, columnType: Column.blob)
+  }, indices: <SchemaIndex>{})
+});
 ''';
 
 /// [SqliteSerializable] **does not** produce code.